First commit
A rather crude version
that can however deal with all three kinds of geometry. Simply vary the parameters p and q to obtain different versions of tessellations.
The output is not optimized but kept as generic as possible. I'm not very skillful with mathematica. If you can improve this worksheet, don't hesitate to do so and I'm happy to update your pull-requests.